test(dsh-provider-usage): 客户端行为抽纯函数断言、react shim 最小类型与纯函数区去 @ts-nocheck (#28)
- 新增 src/client-logic.ts：胶囊文案/状态点/title、D11 隐藏语义、渲染器
  配对键序抽为可导入纯函数（宿主 re-export，client bundle esbuild 内联同源）；
  client/index.ts 与 core.ts 改为组装态调用，smoke-pure.ts 做行为级断言，
  替代对 lib/client.js 的 includes 字符串断言
- react-shim.d.ts 为 useState/useEffect/useCallback/createElement 声明最小
  类型（ReactNode/ReactElement），settings.ts 回填 useState 泛型标注
- 纯函数断言区拆至 test/smoke-pure.ts（无 @ts-nocheck，经 test/tsconfig
  全量类型检查通过）；smoke.ts 仅保留 apply 集成区并复用共享桩
